Created Cart.jsp and updated UpdateInCart.jsp
UpdateInCart.jsp was giving the wrong selecting for wheel. That is now fixed. Cart.jsp will display each item from the InCart table. This may need fixed up based on the tables you've been working on, I'm not sure yet, but if it does, it should be an easy fix